* commit '25c8c198e6abc053fbba19145ed485e31dc5927e':
  Perform a long computation to catch bitmap sizes > 32 bits Bug #7257930

+++ core/java/android/view/View.java
@@ -12811,8 +12811,8 @@ public class View implements Drawable.Callback, KeyEvent.Callback,
             final boolean opaque = drawingCacheBackgroundColor != 0 || isOpaque();
             final boolean use32BitCache = attachInfo != null && attachInfo.mUse32BitDrawingCache;
 
-            final int projectedBitmapSize = width * height * (opaque && !use32BitCache ? 2 : 4);
-            final int drawingCacheSize =
+            final long projectedBitmapSize = width * height * (opaque && !use32BitCache ? 2 : 4);
+            final long drawingCacheSize =
                     ViewConfiguration.get(mContext).getScaledMaximumDrawingCacheSize();
             if (width <= 0 || height <= 0 || projectedBitmapSize > drawingCacheSize) {
                 if (width > 0 && height > 0) {
